js heap growth using timing data https://bugs.webkit.org/show_bug.cgi?id=291274 rdar://148838666
Reviewed by NOBODY (OOPS!).

Previously, we choose m_maxEdenSize (the number of bytes at which we'll trigger gc in a
slow allocation) as a multiple of currentHeapSize (with a lower limit) after Full GC.
We want to adapt to the workload to some degree, allowing ourselves to temporarily use
more memory when we're busy and being careful not to use more memory than necessary
when we're mostly idle.

Adjust m_maxEdenSize on every GC to aim for a desired idealProportionOfTimeSpentDoingGC,
the amount of CPU time spent doing GC divided by the wall time since the last GC.
To ensure that m_maxEdenSize is reasonable, use minHeapSize as the minimum value and
maxEdenSizeLimitMultiplier * minHeapSize as the maximum value.

Since m_maxEdenSize affects how we currently decide to set m_shouldDoFullCollection,
introduce m_maxOldSize to make the decision instead and use a linear function of RAM
usage to choose how much larger than currentHeapSize we should set m_maxOldSize to be
after each Full GC.
(This is similar to how m_maxEdenSize was calculated, but this is a bit simpler.
The idea is to grow more slowly when approaching criticalGCMemoryThreshold.)

To measure the CPU time of each GC cycle, add synchronous timing scopes to the Begin and
End phases as well as the part of Fixpoint that the world is stopped for.
Also accumulate CPU time in the parallel helper threads using the new ParallelCPUTimer,
which takes the sum over the chunks of work done in SlotVisitor::drainFromShared, only
including the maximum set of chunks that don't overlap in time.
ParallelCPUTimer creates RAII objects (CPUTimingScope) to record CPU time across scopes.

Also introduce CPUTime::forCurrentThreadSlow (which uses `thread_info`) as a temporary
workaround for CPUTime::forCurrentThread failing on MacOS due to WCP sandbox.

* Source/JavaScriptCore/heap/Heap.cpp:
(JSC::Heap::Heap):
(JSC::Heap::updateMaxEdenSize):
(JSC::Heap::updateMaxOldSize):
(JSC::Heap::updateAllocationLimits):
* Source/JavaScriptCore/heap/Heap.h:
* Source/JavaScriptCore/runtime/Options.cpp:
(JSC::overrideDefaults):
* Source/JavaScriptCore/runtime/OptionsList.h: